feat: implement dependency walker to streamline transitive dependency discovery
